poc.py: The test harness that demonstrates the proof-of-concept
This patch improves slang.py so that we can express the `validate!` and
`extract!` ideas from the proof-of-concept proposal. There is no logic yet,
just the APIs that we need to express the ideas.

poc.py is a simple test harness that drives the slang API.

We update `make run` to run this demo.

Signed-off-by: Andy Bennett <andyjpb@digital.cabinet-office.gov.uk>