Argument deserialisers required for the `declare_type` directive
A better implementation for the slang.type() deserialiser that looks up
the constructor for the specified type and allocates a new instance of
the appropriate class that can be used to process data from the named
column.

The `declare_type` directive now checks that its arguments look
reasonable, stores the necessary data in the state object and prints
some debug that allows the proof-of-concept to show that things are
working.

The supplied deserialiser assumes a correct implementation of
slang.unescape() but that has not yet been provided. Therefore, metadata
files containing characters that must be escaped will have undefined
behaviour.

Signed-off-by: Andy Bennett <andyjpb@digital.cabinet-office.gov.uk>