Internal data-types to represent the data types that can be found in a spreadsheet cell
The `declare_type` directive allows the user to specify a type for a
particular column of data in a spreadsheet. This patch provides the
internal representation of the types that we know about.

For now we just provide a class hierarchy and a symbolic class for each
distinct type. In time, these classes can be expanded with validators
and deserialisers that can take some raw data that has been extracted
from a spreadsheet cell and parses it into something that we can either
manipulate internally or grind into another serialisation format.

Each class has a quick and dirty __str__ implementation.

In time the contents of this type hierarcy should be drawn from an
external data source, for example a Register, so that the policy and
governanace process for agreeing and defining data types can be separate
from the implementation and mechanism of the code.

For now we just provide the following types that are needed by our
proof-of-concept:

  String  : A collection of unicode codepoints representing human
            readable text.

  GBPxVAT : A number representing a currency value in Stirling. The
            value should exclude Value Added Tax.

  Number  : A number with unspecified units or accuracy.

  Formula : A spreadsheet formula. We expect that the spreadsheet
            formats we are interested in might provide a cached value
            for the cell as well as a serialisation of the formula used
            to calculate it. We don't expect to do much with the formula
            itself although it might be nice to! This is included in the
            proof-of-concept because it highlights the fact that we need
            a way to specify the type of the cached value.

Signed-off-by: Andy Bennett <andyjpb@digital.cabinet-office.gov.uk>