Internal data-types to represent references to Cells and Ranges
The `declare-header` and `declare-data` directives allow the user to
refer to ranges of cells in a spreadsheet. When we deserialise those
specifications we will represent these cells internally as a
RangeReference. RangeReference will, in turn, use a CellReference to
represent each end of the range.

This patch provides initial implementations of CellReference and
RangeReference.

When a CellReference is allocated we calculate and store the cell's row
and column details.

When a RangeReference is allocated we calculate and store the
range'swidth and height details.

Instances of these classes are not intended to be modified after
allocation.

Signed-off-by: Andy Bennett <andyjpb@digital.cabinet-office.gov.uk>