Argument deserialisers required for `declare_header` and `declare_data` directives
Here we provide better implementations for the slang.string() and
slang.range() deserialisers and integrate them into the handlers for the
`declare_header` and `declare_data` directives: `declare_header` expects
a string and a range and `declare_data` expects a range.

These directives now check that their arguments look reasonable, store
the necessary data in the state object and print some debug that allows
the proof-of-concept to show that things are working.

We also provide a bonus slang.cell() deserialiser that can be used to
deserialise a Microsoft Excel style reference to a single cell (for
example "A1"). There are not yet any plans for directives that might use
this but slang.range() could be rewritten in terms of it. It would also
be useful if there was a directive that allowed column types to be
specified directly rather than indirectly via the column header.

slang.range() only supports the basic "A1:B2" format. It can recognise
the basic Microsoft Excel Named Range format but will throw an exception
when it finds one. Ranges scoped by Sheet or Workbook are not yet
recognised.

All of the supplied deserialisers assume a correct implementation of
slang.unescape() but that has not yet been provided. Therefore, metadata
files containing characters that must be escaped will have undefined
behaviour.

Signed-off-by: Andy Bennett <andyjpb@digital.cabinet-office.gov.uk>