Initial abstraction around the spreadsheet itself
Here we provide a wrapper for a Cell inside an ODF document.

A single slang metadata description should be able to extract data from
a range of Spreadsheet formats. For the proof-of-concept we will
implement extraction from ODF files using the ODFPY library.

This patch specified a class hierarchy that (hopefully) allows Cells
from other Spreadsheet formats in the future. The interface provides a
way to get the raw data from a cell as well as type information about
the cell. The factorisation of this may have to be adjusted a number of
times as more spreadsheet formats are supported.

We expect that the data returned from the accessors be represented in an
intermediate format that can be turned into the higher level domain
types. For now we don't go much further than ensuring that the data is
a valid unicode string. The metadata around the cell type is represented
as a series of predicates. OdfCell.test_attribute() is intended to be an
internal procedure, not part of the API.

We also provide a way to get extended metadata about formulae and
currencies but we don't make any attempt to define an intermediate
format for this data.

Signed-off-by: Andy Bennett <andyjpb@digital.cabinet-office.gov.uk>