poc-json.py
A version of poc.py that outputs the data extracted from the spreadsheet
as a JSON document.

This is just to demonstrate the idea that once we can extract and
validate the data from the spreadsheet, we can re-render it as something
more useful. This could easily be adapted to, for example, emit the data
as a CSV file with a CSV on the Web metadata file.

Signed-off-by: Andy Bennett <andyjpb@digital.cabinet-office.gov.uk>