A quick and dirty spreadsheet decoder
Extract data from the spreadsheet using the metadata.

Find the header and link each column up with a constructor procedure for
the appropraite slang Type.

Find the data (by assuming that it is in the first sheet of the
workbook), extract each row and apply the correct Type constructor to
each Cell, returning an array.

This processes the entire data range all at once and we don't go to any
special lengths to be fast or efficient.

We try to support data expressed in rows with column headers as well as
data expressed in columns with row headers. However, the latter doesn't
work yet.

Signed-off-by: Andy Bennett <andyjpb@digital.cabinet-office.gov.uk>