Simple, high level, validation of the metadata
This patch provides a routine that validates a metadata description as a
whole after all the directives have been processed.

Each directive performs its own checks but becase of the ordering
constraints and the way that directives interact, it is possible for a
metadata description to be invalid.

state.validate() checks that we have most of the stuff we need but the
resulting metadata description may still fail when trying to extract
data from arbitrary spreadsheets.

Signed-off-by: Andy Bennett <andyjpb@digital.cabinet-office.gov.uk>